refactor the layout for database access and models
