+++ uprn_mangle/backend/config/__init__.py
@@ -1,5 +1,5 @@
 """Configuration module."""
 
-from .settings import Settings
+from .settings import Settings, get_settings
 
-__all__ = ["Settings"]
+__all__ = ["Settings", "get_settings"]
