+++ uprn_mangle/backend/database/db.py
@@ -1,0 +1,36 @@
+"""Configure the Database."""
+
+from sqlalchemy import MetaData, create_engine
+from sqlalchemy.orm import DeclarativeBase, sessionmaker
+
+from uprn_mangle.backend.config import get_settings
+
+settings = get_settings()
+
+DATABASE_URL = (
+    f"postgresql://{settings.db_user}:"
+    f"{settings.db_password}@"
+    f"{settings.db_host}:"
+    f"{settings.db_port}/"
+    f"{settings.db_name}"
+)
+
+
+class Base(DeclarativeBase):
+    """Base class for SQLAlchemy models."""
+
+    metadata = MetaData(
+        naming_convention={
+            "ix": "ix_%(column_0_label)s",
+            "uq": "uq_%(table_name)s_%(column_0_name)s",
+            "ck": "ck_%(table_name)s_%(constraint_name)s",
+            "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
+            "pk": "pk_%(table_name)s",
+        }
+    )
+
+
+sync_engine = create_engine(DATABASE_URL, echo=False)
+session_local = sessionmaker(
+    autocommit=False, autoflush=False, bind=sync_engine
+)
