+++ uprn_mangle/backend/models.py
@@ -1,23 +1,11 @@
 """Pydantic and SQLAlchemy models for the address table."""
 
 from pydantic import BaseModel
-from sqlalchemy import BigInteger, Column, Float, Index, MetaData, String
+from sqlalchemy import BigInteger, Column, Float, Index, String
 from sqlalchemy.dialects.postgresql import TSVECTOR
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
+from sqlalchemy.orm import Mapped, mapped_column
 
-
-class Base(DeclarativeBase):
-    """Base class for SQLAlchemy models."""
-
-    metadata = MetaData(
-        naming_convention={
-            "ix": "ix_%(column_0_label)s",
-            "uq": "uq_%(table_name)s_%(column_0_name)s",
-            "ck": "ck_%(table_name)s_%(constraint_name)s",
-            "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
-            "pk": "pk_%(table_name)s",
-        }
-    )
+from uprn_mangle.backend.database import Base
 
 
 class Address(Base):
